requirements.txt
bcrypt==4.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.0.3
fonttools==4.54.1
func_timeout==4.3.5
future==1.0.0
hnmp==0.5.1
idna==3.10
igraph==0.10.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
n2g==0.3.3
netmiko==4.4.0
networkx==3.2.1
ntc_templates==7.4.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
paramiko==3.5.0
pillow==11.0.0
ply==3.11
pyasn1==0.4.8
pycparser==2.22
pycryptodomex==3.21.0
PyNaCl==1.5.0
pyparsing==3.2.0
pyserial==3.5
pysmi==1.5.9
pysnmp==4.4.12
python-dateutil==2.9.0.post0
python-igraph==0.10.8
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
scp==0.15.0
six==1.16.0
tabulate==0.9.0
textfsm==1.1.3
texttable==1.7.0
ttp==0.9.5
ttp-templates==0.3.7
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.20.2
